requirements.txt
Django==4.1.*
djangorestframework==3.14.*
requests==2.28.*
environs==9.5.*
python-telegram-bot==13.*